Create campaign prototype
A *working prototype* of the dapp. Right now it does only a couple of things,
it's missing tests, and it doesn't rely on any modern framework, but that's a
proof of concept that things work as expected.

The dapp needs to interact with an Ethereum node through `web3.js`. It can be
used from Mist or from any browser that has installed the MetaMask plugin.

==== Current account and current block info

The first section of the dapp includes information about the current account and
the current block. The dapp listens to new block events in the blockchain and
updates the first section of the page.

===== Before connecting an account

![](https://gist.githubusercontent.com/potomak/bcc89e435a29f14f2405ca53c9d18258/raw/d1567240d741fcf85405c81857b04826d2dd7de7/Screen%2520Shot%25202017-09-06%2520at%252012.00.13%2520AM.png)

===== After connecting an account

![](https://gist.githubusercontent.com/potomak/bcc89e435a29f14f2405ca53c9d18258/raw/d1567240d741fcf85405c81857b04826d2dd7de7/Screen%2520Shot%25202017-09-06%2520at%252012.00.54%2520AM.png)

==== Creating a new campaign

The second section includes a form to create a new campaign.

![](https://gist.githubusercontent.com/potomak/bcc89e435a29f14f2405ca53c9d18258/raw/d1567240d741fcf85405c81857b04826d2dd7de7/Screen%2520Shot%25202017-09-06%2520at%252012.01.15%2520AM.png)

"Create campaign" will send a call to `createCampaign` on the issuehunter
contract deployed on rinkeby.io. The call triggers a dialog window that asks for
a transaction confirmation.

![](https://gist.githubusercontent.com/potomak/bcc89e435a29f14f2405ca53c9d18258/raw/d1567240d741fcf85405c81857b04826d2dd7de7/Screen%2520Shot%25202017-09-06%2520at%252012.01.27%2520AM.png)

After the user confirmed the transaction the dapp will update the interface by
reacting to new `CampaignCreated` events.

===== Before a `CampaignCreated` happens

![](https://gist.githubusercontent.com/potomak/bcc89e435a29f14f2405ca53c9d18258/raw/d1567240d741fcf85405c81857b04826d2dd7de7/Screen%2520Shot%25202017-09-06%2520at%252012.01.43%2520AM.png)

===== After a `CampaignCreated` happened

![](https://gist.githubusercontent.com/potomak/bcc89e435a29f14f2405ca53c9d18258/raw/d1567240d741fcf85405c81857b04826d2dd7de7/Screen%2520Shot%25202017-09-06%2520at%252012.02.08%2520AM.png)

See also:

* https://gist.github.com/potomak/bcc89e435a29f14f2405ca53c9d18258#file-readme-md
* https://github.com/issuehunter/dapp/issues/4